pyproject.toml
[project]
name = "microc"
version = "0.1.0"
description = "Interpretador MicroC em Python"
authors = [{ name = "Seu Nome" }]
requires-python = ">=3.10"
dependencies = [
    "ipdb>=0.13.13",
 "lark-parser>=0.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-timeout>=2.4.0",
]

[project.scripts]
microc = "MicroC.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:lark.utils",
]
